Refactor reviews and reading log features with improved filtering architecture
